# openapi.yaml
openapi: 3.0.0
info:
  title: Kallipolis ZK Security API
  version: 3.0.0
  description: |
    AI-Powered Security Layer for Polygon AggLayer.
    Features:
    - Real-time MEV detection
    - Smart contract audit (Slither, Mythril, Gemini)
    - Bridge verification
    - ZK proof generation
    - Formal verification
servers:
  - url: https://api.kallipolis.ai/v1
    description: Production server
  - url: http://localhost:3000/api/v1
    description: Development server
paths:
...
  /audit/analyze:
    post:
      summary: Smart contract security audit
      description: Analyzes Solidity source code, bytecode, or contract address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                source_code:
                  type: string
                  description: Solidity source code
                bytecode:
                  type: string
                  description: Contract bytecode (hex)
                contract_address:
                  type: string
                  description: Contract address on Polygon
      responses:
        '200':
          description: Analysis results
          content:
            application/json:
              schema:
                type: object
                properties:
                  audit_id:
                    type: string
                    example: "AUDIT-CBQZ5L9O"
                  risk_score:
                    type: number
                    minimum: 0
                    maximum: 100
                    example: 96
                  overall_status:
                    type: string
                    enum: [SECURE, REVIEW_NEEDED, VULNERABLE]
                  vulnerabilities:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        title:
                          type: string
                        severity:
                          type: string
                          enum: [CRITICAL, HIGH, MEDIUM, LOW]
                        category:
                          type: string
                        line:
                          type: integer
                        description:
                          type: string
                  formal_verification_passed:
                    type: boolean
                  gas_optimizations:
                    type: array
                    items:
                      type: string
                  execution_time_ms:
                    type: number
...
  /bridge/verify:
    post:
      summary: Bridge Merkle root verification
      description: Validates Merkle root for LxLy bridges
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                merkle_root:
                  type: string
                chain_id:
                  type: integer
      responses:
        '200':
          description: Verification result
          content:
            application/json:
              schema:
                type: object
                properties:
                  verified:
                    type: boolean
                  checks:
                    type: object
                    properties:
                      format:
                        type: boolean
                      rpc:
                        type: object
                      zk_proof:
                        type: boolean
...